Plugins: Unwrap - fix initial state
Initialize m_prev and restore initialized state after calculating init
sample.

This is for completeness with the previous commit, perhaps a moot point
because there appears to be no language-side class definition of the
Unwrap UGen. (It should probably be either updated or removed
altogether because on casual inspection this implementation doesn’t
allow modulatable arguments.)