plugins: TWindex_Ctor rewritten to fix initialization sample and first output sample.
There was an an erroneous assumption that a trigger will
always happen on the first sample (thereby always generating an initial
random value), because `unit->m_trig = -1.f;`), which isn't the case if
first trig input is `<=0`, and because `m_index` is initialized to `0`.
So if first trig input is `<=0`, the first output is _always 0_. This
change fixes that, achieving both a guaranteed random value on the
first sample, and also that it will match the initialization sample.