[plugins] Klang - fix init sample.
Previously the init sample was correct, but y1 was set to the same
value as y0, and y2 was set to the value of y1. This update makes y0
independent, properly sets y1 and y2 such that the both the init sample
and the first sample will be equal and in their proper place in time.