[plugins] Lag, Lag2, Lag3, LagUD, Lag2UD, Lag3UD: y1=y0 on init
`y1` restored to be equal to the first input sample. This runs counter to the general rule of setting assuming silence preceding the first sample, but the `Lag` family of UGens are commonly used for control signals, which cannot be assumed to be bound between [-1,1], so fading in (lagging) from a previous value of 0 is not ideal.

As a consequence some a-rate inputs don't have ideal behavior, such as an impulse train being smoothed (first impulse sample is full while proceeding are attenuated/filtered). Future change could set different `y1` values for `kr` and `ar` inputs, but that needs discussion.